==> 1_DFWGCNA.Rmd <==
---
title: "lIMMA and WGCNA analysis for LPS response gene matrix"
output: html_notebook
date: "2023-11-12"
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Cmd+Shift+Enter*. 

```{r}
library(dplyr)
library(tibble)
library(limma)
library(ggplot2)
```

# load gene expression data and wgcna clusters
```{r}
# gene info includes the gene name, TLR4 genes, and WGCNA clusters ('mergedColors2A.diff')
# we did not rerun WGCNA because the clusters may change for each time clustering.
geneinfo = readRDS('data/geneinfo.rds')
load('data/LPSgeneexpr.RData')
# normlps: log2 LPS gene expression
# normbase: log2 baseline gene expression
# diffexpsfiter: LPS response matrix (log2LPS - log2Base),the rows are entrez id
# diffexps_hgnc_f : LPS response matrix (log2LPS - log2Base),  the rows are gene symbol
allgenesDFlpsfilter = geneinfo


```


# limma logFC FOR diffexpsfiter
```{r}
DER <- topTable(eBayes(lmFit(diffexpsfiter)), number=nrow(diffexpsfiter)) %>% rownames_to_column(var = 'genes')
#ggplot(DER, aes(x = logFC, y=-log10(adj.P.Val)))+geom_point()
allgenesDFlpsfilter = allgenesDFlpsfilter %>% left_join(DER, by = 'genes')
```

# save to 'DFWGCNA.rds'
```{r}
saveRDS(allgenesDFlpsfilter, 'data/DFWGCNA.rds')
```






Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Cmd+Option+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Cmd+Shift+K* to preview the HTML file). 

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.
